상품 리스트 조회 #3

+++ src/main/java/com/example/backoffice/product/dto/GetProductResponse.java
@@ -3,6 +3,7 @@ package com.example.backoffice.product.dto;
 import com.example.backoffice.admin.entity.Administrator;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -10,14 +11,14 @@ public class GetProductResponse {
     private final Long id;
     private final String name;
     private final String category;
-    private final Long price;
+    private final BigDecimal price;
     private final Long stock;
     private final String status;
     private final LocalDateTime createdAt;
     private final LocalDateTime modifiedAt;
     private final Administrator admin;
 
-    public GetProductResponse(Long id, String name, String category, Long price, Long stock, String status, LocalDateTime createdAt, LocalDateTime modifiedAt, Administrator admin) {
+    public GetProductResponse(Long id, String name, String category, BigDecimal price, Long stock, String status, LocalDateTime createdAt, LocalDateTime modifiedAt, Administrator admin) {
         this.id = id;
         this.name = name;
         this.category = category;
